Simplify handling of attached debuggers.
Instead of having to carefully update the reader to point through the
debugger, simply create a single ObservableReader/Writer on startup
and point the debuggers at them.  Whenever the stream changes, we just
change the underlying stream of those classes.  This simplifies all of
the debuggers, and simplifies the logic for creating reader/writer
streams.