Improve certificate handling.
If a secure connection isn't required, pass this along to ServerTrustManager
and XMPPSSLSocketFactory.  In that case, don't do certificate checks from
checkServerTrusted; if we throw a failure, the connection is aborted with
an exception.  Instead, if the security setting isn't "required", do it
from handshakeCompleted and remember the result for use elsewhere.