Refactor transport handling into its own class, XMPPStream.
The transport layer is the TCP connection and things specific to
it: compression and TLS.  This encapsulates stream negotiation,
so it can be replaced.

This also improves PacketReader and PacketWriter abstraction.  Both
of these classes control threads, with somewhat complex interactions
with the connection, so having these as clearly defined and separated
as possible is important.

This was too involved to do piecemeal, so a list of notable results:

- The basic interface for reading packets is XMPPStream.readPacket, which
blocks and returns an Element.  The use of XmlPullParser is abstracted;
a whole stanza will be read and returned.
 - This also means that we can expose Elements to higher-level code, which
   is a much saner way to read XML trees, and we can expose the stanza
   directly from Packet to aid extensions; there's no need to do all
   parsing from PacketReader.
- TLS negotiation is improved.  The TLS object will be initialized locally
in advance, rather than when TLS is requested.  This way, if TLS is offered
but we don't support it locally, we don't request it and then fail; we just
don't request it in the first place.
- Keepalives are a transport-level issue, and are moved to XMPPStreamTCP.
Other protocols handle them differently.
- Stream negotiation is handled by processInitializationPacket, in a
self-contained way.  The user of XMPPStream sees no packets before the
first <feature/> where we're not requesting TLS or compression.
- If TLS compression is enabled, XMPP compression will not be used.
- PacketWriter no longer has direct access to PacketReader.
- PacketReader and PacketWriter startup and shutdown are now better behaved,
shutting down more reliably and propagating errors through startup() more
reliably.
- Arbitrary "wait 150ms after shutting down" removed; cleanup() now not
return until the thread is fully shut down.