Separate reader/write observer class from events.
ObservableReader/ObservableWriter did two things: intercepted reads
and writes, and delivered events.  Separate this into two classes (each,
for now): the read/write filter and the event.

Thus,

  event = new WriteEvent();
  obsWriter = ObservableWriter(writer);
  obsWriter.setWriteEvent(event);

This allows the event to be separated from the writer.  We can create
the read/write events on initialization, and create the debugger
class immediately, passing them the (currently inactive) read/write
events.  Then, when we later connect to a server, we attach the events
to whatever low-level stream is actually used.  In the case of BOSH,
no ObservableReader/ObservableWriter is used at all; it just delivers
events directly.

These event classes could be separated into a different class, and
should probably be merged into a single event class instead of two
separate but logically identical event classes.