Improve supportSASLMechanism().
If the index given is out of bounds, clamp it to the last entry, so
supportSASLMechanism(mech, 9999) will always add a mechanism at the
lowest-priority.

Register internal mechanisms lowest-priority first, passing 0 each time.
This allows mechanisms to be disabled without having to reorder all
of the mechanisms ahead of it.