Improve SASL mechanism searching.
SASL mechanisms were throwing NPE when SASL clients for a mechanism
couldn't be found.  Handle this cleanly: if a mechanism isn't actually
supported, move on to the next one.

Only fall back on legacy authentication if no working SASL mechanisms
were found.  Make this fallback work for the CallbackHandler case.