Run listeners and queue collectors from the same thread.
Collectors are asynchronous; there was no guaranteed of whether collectors
would run before or after listeners.

Listeners are synchronous in their own thread, so they were guaranteed to
run one at a time in packet order, but with no connection to collectors.

Now, collectors are queued from the listener thread.  This means that
they're still asynchronous, but collectors for a packet are guaranteed
never to run until listeners for a packet.

This allows using a collector to block until listeners have run.