Refactor Roster and RosterGroup handling.
Upstream r11634 broke roster handling: making changes to rosters waits
until the roster push to reflect them locally, which makes the roster
API racy and hard to work with.  After making a change to the roster
you need to wait until the roster push before making another, and
there's no way to do that with XMPP: roster pushes aren't tagged with
the ID of the command that caused them.

Refactor RosterPacketListener to make it less monolithic, in order
to make this easier to fix.

addedEntries, updatedEntries and deletedEntries are now persistant.
Changes made locally will be queued in these sets; they'll still only
be sent in a notification when a roster push arrives.

When changes are made locally, reflect them locally as soon as the
command is successful.  Then, when the roster push arrives, we update
with the authoritative data.

There's one tricky part: the roster push happens asynchronously.  We
need to make sure that our local update always happens after the
command is successful, and always before the roster push is processed.
Do this by performing the local update in a PacketListener.  PacketListeners
are run from a single worker thread; they're guaranteed to happen one
at a time and in packet order, so this will always be run before
the roster push.

This also uses a related change in PacketReader: we're guaranteed to
only receive the packet in the PacketCollector *after* PacketListeners
have run, so we'll never remove the collector before it's had a chance
to run.

This all makes the API much more predictable and deterministic, and
helps fix a bunch of roster tests.