RosterSmackTest fixups and compile fix.
If A subscribes to B, A is added to B's roster to register the "from"
subscription.  If A then unsubscribes, B's roster entry is set to "none"
but A is not actually removed from B's roster.  So, B's roster entry
count is 1, not 0.  (Testing with Prosidy; I can't find anything about
this in the spec.)

Fix some test races.  Use a Timeout class to make the timeout loops
clearer.