Simplify and fix up PacketReaderTest.testFiltersRemotion.
Only test in one direction, so it's easier to diagnose when this test
doesn't work.  Only run more complicated tests when actually needed,
and only after running the simpler cases first.

Fix race conditions: this was sleeping arbitrarily for listeners to
run, which wasn't always working.  Use a collector to wait properly.