Handle disconnections gracefully.
When sending a disconnect, send the message and wait for the server to
close the connection before returning.  This ensures that the server
receives the termination message (if possible), that any packets the
server has buffered are flushed, and that the server has a chance to
send any termination messages in response.

This also refactors XMPPStream to stop exposing the output as a Writer,
which allows XMPPStreamBOSH to be simplified; it previously exposed
itself as a Writer to to behave like XMPPStreamTCP.  Now, we just take
packet data directly.

This also avoids exposing the XMPPStream directly to PacketWriter.