Start to fix conflation of XMPPError and XMPPException.
The mixing of "message" in these classes is confusing.  Use XMPPError to
represent an actual error on the wire: a message sent to or from the
server in an error packet.  Use XMPPException's message as normal
exception messages are used.

This means that it may make sense to propagate an XMPPError's message
to XMPPException, but it rarely makes sense to propagate an exception
message to XMPPError.