Refactor SASL authentication.
PacketReader shouldn't need to care about SASL.  Treat SASL roughly like
anything else, using PacketCollector to wait for responses.

This means:

- PacketReader doesn't reset the stream on <success/>.  SASLAuthentication
does this now.
- PacketReader doesn't have to parse SASL-specific features.
- The SASLAuthentication object doesn't need to be kept around forever.
Create it when we need it, and then discard it.

This uses a new packet type, ReceivedPacket.  This is a generic packet
type, containing a parsed DOM node.  ReceivedPacketFilter can be used
to filter on ReceivedPackets with a specific localName and namespace.
This packet type can be used for packets not widely-used enough to
warrant setting up entire Packet types--primarily <features/> and SASL
responses.