Fix bugs in ServerTrustManager.
Hostname matching was broken: *.bar.com would match foobar.com.  Rewrite
this based on org.apache.http.

Remove date checking; CertPathValidator does it already.

Fix NPE if javax.net.ssl.trustStore is unset.

Remove confusing mention of the subjectAltName "xmpp" type; that's
used for client-to-server and server-to-server auth, and is irrelevant
here where we're dealing with server-to-client auth.