Fix race condition during connect.
When <features> is received, PacketReader queues it to be sent to
listeners, and returns from startup.  There's no guarantee that
it'll be returned from listeners before returning; this causes
a race, since login() might be called before the listener is run.

Fix and simplify this: use a collector to grab the initial <features>
during connection.