Add a SCRAM-SHA-1 SASL implementation.
This is required by XMPP.

This is implemented directly, rather than as a javax.security.sasl
provider.  This moves towards being able to function without that
package.  It's not available on Android, and the Java brokenness
that makes us need to rename the javax package is ugly, so it'd
be nice to be able to function without it eventually.  The only
SASL mechanisms most people need are SCRAM-SHA-1, DIGEST-MD5 and
PLAIN anyway.