Fix blocking for PacketCollectors from within PacketListeners.
This was a side-effect of moving collectors to the executor thread.
This broke testSendAndAcceptRoster, because the following happens:

- testSendAndAcceptRoster sends a packet.
- rosterExchangeListener receives it, and calls Roster.createEntry.
- Roster.createEntry creates a listener, sends the roster set packet,
and then blocks on a collector waiting for the response.

Since createEntry was already running in the listener thread, neither
listeners nor collectors were running, so it blocked indefinitely.

This is something that generally shouldn't be done anyway: if you
block within a listener, no other listeners can run.  However, API
restrictions like "can't run Roster.createEntry from within a listener"
are better avoided, so let's make this work.  Fix this with the following:

- Revert 22442b7d, which moved collector queueing into the listener
thread.  Collectors are again notified from the PacketReader thread.
- Add a new subclass of PacketListener, SynchronousPacketListener.
Listeners of this class are run from the PacketReader thread,
immediately before notifying collectors.

Synchronous packet listeners allow processing responses to packets
in the order they happen, as is needed for roster changes (to store
changes while guaranteeing any following roster push always happens
afterwards).  However, they block the PacketReader thread and must
never block on a collector.  They should only update data based on
a result packet.

(SynchronousPacketListener should really just be an "isSynchronous"
method on PacketListener, but that would require changing every
implementation of PacketListener.  PacketListener should be changed
to an abstract class.)