Overhaul XMPPStreamTCP locking, and switch to a callback model.
Object monitors are insufficient for nontrivial locking.  Switch to a
lock and condition, so the lock status can be validated at each step,
guaranteeing that we're not locked when we don't expect it or vice
versa.

Switch to a callback model for stream events, to eliminate the PacketReader
thread.

(This is a large refactoring; they go together and aren't worth trying
to pull apart.)