If isRosterLoadedAtLogin is false, disable the roster.
This code path was problematic.  It didn't handle locking properly;
if multiple threads called getRoster in parallel, reload() could
be called multiple times.  It didn't handle timeouts at all, returning
a bogus roster after a timeout instead of throwing an error.

Just disable the roster if isRosterLoadedAtLogin is disabled.  Either
load the roster if you're going to use it, or don't load it if you're
not.  If rosters are taking too long to download, then roster versioning
should be used.

The "synchronize against login" code path still doesn't make sense.