Fix connection timeouts cause all transports to be cancelled.
cancel() should cancel the entire connection process, but a transport
timeout should move on to the next connection attempt.

This also fixes discovery timeouts.  When discovery for a service
times out, it should attempt a default connection, which is the
same thing that happens when discovery fails.  To do this,
XMPPStream.getDefaultConnectData() is separated from getConnectData;
this function never fails or blocks, and can be used as a fallback
when getConnectData times out.