Initial pass at fixing up proxy connection.
Proxy connection needs to use async DNS lookups and to allow cancellation,
just like a regular connection.  Using a SocketFactory isn't appropriate
for this.

Create a new SocketConnectorFactory, which will handle async DNS,
cancellation and proxy connection.

This is a first pass that doesn't implement most of this; it just
reorganizes the proxy classes into this class structure to get them
ready for it.  Most of the changes in the proxy classes is just
reindentation.