Change ConnectionListener to a class and eliminate AbstractConnectionListener.
Interfaces are for published API interfaces where the interface itself
is part of the exported API.  Methods can't easily be added to interfaces,
because it's backwards-incompatible: every user of the interface needs
to update code to implement the new method.  ConnectionListener isn't
exporting an API, it's importing callbacks; new callbacks can be added
without user code ever needing to be changed.  Everyone should just be
using AbstractConnectionListener, but even within the library itself
this isn't happening.

Just change ConnectionListener to a class.  This is backwards-incompatible,
but no more so than adding a method to ConnectionListener.  It also
eliminates the unnecessary duplicateion between ConnectionListener
and AbstractConnectionListener.