Don't swallow Exception in Socks5Proxy.
Wait: a socks proxy server?

I may just rip this out.  It seems to only be needed for the server
side; this code should focus on the client side.  Also, whoever wrote
XEP-0065 can't have much real experience with production servers: it's
completely impractical for any production service to handle file
transfers via proxying.

This should be scrapped and replaced with STUN/TURN, which allows for
direct peer-to-peer connections even when both parties are behind NAT.